\name{gama05}
\docType{data}
\alias{gama05}
\title{Goosenest Adaptive Management Area Data from 2005}
\description{
  This data file has 20 observations on plots from four different
  treatments on the Goosenest Adaptive Management Area (Ritchie 2005).
  It includes qmd, ba, and tpa (English units) for each plot.
  The file also includes a plot number called UNIT and a treatment
  identifier (character string) for each plot. when using gama05, use.metric=FALSE.
}
\usage{gama05}

\references{
    Ritchie, M.W. (2005) Ecological research at the Goosenest Adaptive
      Management Area in northeastern California. Gen. Tech. Rep.
      PSW-192. USDA For. Serv. Pac. SW Res. Stn. 120p.

}
\author{
Martin W. Ritchie, Research Forester, US Forest Service, Pacific Southwest Research Station.
  Email: \email{martin.ritchie@usda.gov}
}
\keyword{ datasets }
