\name{gama05}
\docType{data}
\alias{gama05}
\title{Goosenest Adaptive Management Area Data from 2005}
\description{
  This data file has 20 observations on plots from four different
  treatments. It includes qmd, ba, and tpa (English units) for each plot.
  The file also includes a plot number called UNIT and a treatment
  identifier (character string) for each plot. when using gama05, use.metric=FALSE.
}
\usage{gama05}

\references{
    Ritchie, M.W. (2005) Ecological research at the Goosenest Adaptive
    Management Area in northeastern California. Gen. Tech. Rep.
    PSW-192. USDA For. Serv. Pac. SW Res. Stn. 120p.

    Drew, T.J. and J.W. Flewelling (1979). Stand density management: an alternative
    approach and its application to Douglas-fir plantations. For. Sci. 25(3):518-532

    Cochran, P.H. (1992) Stocking Levels and Underlying Assumptions for Uneven-Aged
    Ponderosa Pine Stands. Res. Note PNW-509. USDA For. Serv. Pac. Northwest For.
    and Range Exp. Stn. 10p

    Long, J.N. and Shaw, J.D. (2005) A density management diagram for even-aged ponderosa
    pine stands. West. J. Appl. For. 20(4):205-215.

    Reineke, L.H. (1933) Perfecting a stand density index for even-aged forests.
    J. Ag. Res. 46:627-638.
}
\author{
Martin W. Ritchie, Research Forester, US Forest Service, Pacific Southwest Research Station.
  Email: \email{mritchie@usda.gov}
}
\keyword{ datasets }
